upating message

+++ a^b.c
@@ -3,16 +3,16 @@
 #include<stdio.h>
 void main ()
 {
-int a, b, num1=1, val=1;
+double a, b, num1=1, val=1;
 printf("Enter the base number :\n");
-scanf("%d",&a);
+scanf("%lf",&a);
 printf("Enter the power :\n");
-scanf("%d",&b);
+scanf("%lf",&b);
 while (num1 <= b)
     {
         val = val*a;
         num1++;
     }
-printf("The value of %d to the power %d is : %d\n", a, b, val);    
+printf("The value of %lf to the power %lf is : %lf\n", a, b, val);    
 
 }
